+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+### Added
+- Get cluster labels method to InstanceClient class.
 ### Fixed
 - Possible password leak in LocalExecutor.